exo/requirements/base.txt
django>1.9,<1.10
django-environ==0.4.1
Pillow==3.4.2
django-model-utils==2.6
awesome-slugify==1.6.5
pyyaml
django_autoslug
django-badges==0.2.0